requirements.txt
numpy==1.21.5
pandas==1.3.5
neomodel==4.0.8
biopython==1.79
tqdm==4.62.3
pytz==2021.3
networkx==2.6.3
requests==2.27.1
w3lib==1.22.0
Whoosh==2.7.4
diskcache==5.4.0
itemloaders==1.0.4
Scrapy==2.6.0
itemadapter==0.4.0
openpyxl==3.0.5
xlrd==1.2.0